requirements.txt
openpyxl~=3.1.5
pandas~=2.2.3
pillow~=10.4.0